+++ tests/corpus/raw-baseline/arm64_O1/xxhash32.c
@@ -27,19 +27,19 @@ uint32_t sym__xxhash32(uint64_t X0_0, uint64_t X1_0, uint32_t W2_0)
                 uint64_t tmp_7c00_2 = X12_5;
                 uint32_t tmp_24280_2 = *(uint32_t*)tmp_7c00_2;
                 uint32_t tmp_24480_2 = ((uint32_t*)tmp_7c00_2)[1];
-                uint32_t tmp_28e80_2 = (uint32_t)X15_1 - tmp_24280_2 * 0x7a143589;
-                uint32_t tmp_28e80_3 = (uint32_t)X16_1 - tmp_24480_2 * 0x7a143589;
+                uint32_t tmp_28e80_2 = X15_1 - tmp_24280_2 * 0x7a143589;
+                uint32_t tmp_28e80_3 = X16_1 - tmp_24480_2 * 0x7a143589;
                 uint32_t* tmp_7b80_2 = (uint32_t*)(X12_5 + 8);
                 uint32_t tmp_24280_3 = *tmp_7b80_2;
                 uint32_t tmp_24480_3 = tmp_7b80_2[1];
-                uint32_t tmp_28e80_4 = (uint32_t)X2_0_2 - tmp_24280_3 * 0x7a143589;
+                uint32_t tmp_28e80_4 = X2_0_2 - tmp_24280_3 * 0x7a143589;
                 tmp_2b380_2 = (tmp_28e80_2 >> 19 | tmp_28e80_2 << 13) * 0x9e3779b1;
                 X15_1 = (uint32_t)tmp_2b380_2;
                 tmp_2b380_3 = (tmp_28e80_3 >> 19 | tmp_28e80_3 << 13) * 0x9e3779b1;
                 X16_1 = (uint32_t)tmp_2b380_3;
                 tmp_2b380_4 = (tmp_28e80_4 >> 19 | tmp_28e80_4 << 13) * 0x9e3779b1;
                 X2_0_2 = (uint32_t)tmp_2b380_4;
-                uint32_t tmp_28e80_5 = (uint32_t)X14_1 - tmp_24480_3 * 0x7a143589;
+                uint32_t tmp_28e80_5 = X14_1 - tmp_24480_3 * 0x7a143589;
                 tmp_2b380_5 = (tmp_28e80_5 >> 19 | tmp_28e80_5 << 13) * 0x9e3779b1;
                 X14_1 = (uint32_t)tmp_2b380_5;
                 X12_5 += 16;
@@ -52,7 +52,7 @@ uint32_t sym__xxhash32(uint64_t X0_0, uint64_t X1_0, uint32_t W2_0)
         {
             uint32_t X14_11;
             uint64_t X15_10;
-            X14_11 = (uint32_t)((uint32_t)X1_0 + (uint32_t)X14_8);
+            X14_11 = (uint32_t)((uint32_t)X1_0 + X14_8);
             uint64_t X15_9 = X12_5 + 4;
             if (X15_9 <= X11_1) {
                 for (; ; ) {
@@ -60,7 +60,7 @@ uint32_t sym__xxhash32(uint64_t X0_0, uint64_t X1_0, uint32_t W2_0)
                     uint64_t tmp_7400_2 = X12_5;
                     X12_5 += 8;
                     uint32_t tmp_24e00_2 = *(uint32_t*)tmp_7400_2;
-                    uint32_t tmp_28e80_8 = (uint32_t)X14_11 - tmp_24e00_2 * 0x3d4d51c3;
+                    uint32_t tmp_28e80_8 = X14_11 - tmp_24e00_2 * 0x3d4d51c3;
                     X14_11 = (uint32_t)((tmp_28e80_8 >> 15 | tmp_28e80_8 << 17) * 0x27d4eb2f);
                     uint8_t ZR_7 = X12_5 == X11_1;
                     uint8_t CY_7 = X11_1 <= X12_5;
@@ -80,7 +80,7 @@ uint32_t sym__xxhash32(uint64_t X0_0, uint64_t X1_0, uint32_t W2_0)
                         uint64_t tmp_7400_5 = X15_10;
                         X15_10++;
                         uint8_t tmp_25400_2 = *(uint8_t*)tmp_7400_5;
-                        uint32_t tmp_28e80_11 = (uint32_t)X14_11 + (uint32_t)tmp_25400_2 * 0x165667b1;
+                        uint32_t tmp_28e80_11 = X14_11 + (uint32_t)tmp_25400_2 * 0x165667b1;
                         X14_11 = (uint32_t)((tmp_28e80_11 >> 21 | tmp_28e80_11 << 11) * 0x9e3779b1);
                         uint8_t TMPZR_24 = X11_3 == 1;
                         X11_3--;
